Reorganize formVariant files a bit
