+++ src/features/kartleggingssporsmal/form/KartleggingssporsmalForm.tsx
@@ -6,9 +6,9 @@ import { revalidateLogic } from "@tanstack/form-core";
 import { useState } from "react";
 import { logTaxonomyEvent } from "@/analytics/logTaxonomyEvent";
 import { kartleggingssporsmalFormFields } from "@/forms/kartleggingssporsmal/fieldDefinitions/allFields";
-import type { FormVariant } from "@/forms/kartleggingssporsmal/formVariants/formVariants";
-import { getFieldIdsToIncludeInForm } from "@/forms/kartleggingssporsmal/formVariants/getFieldsToIncludeInForm/getFieldsToIncludeInForm";
-import { getSchemaForVariant } from "@/forms/kartleggingssporsmal/formVariants/SchemaForVariant";
+import { getFieldIdsToIncludeInForm } from "@/forms/kartleggingssporsmal/formVariants/getFieldsToIncludeInForm";
+import type { FormVariant } from "@/forms/kartleggingssporsmal/formVariants/types/FormVariant";
+import { getSchemaForVariant } from "@/forms/kartleggingssporsmal/formVariants/types/SchemaForVariant";
 import { useAppForm } from "@/hooks/form";
 import { submitFormAction } from "@/services/meroppfolging/actions/submitFormAction";
 import type { KartleggingssporsmalFormResponse } from "@/services/meroppfolging/schemas/requestsAndResponses";
