+++ src/forms/kartleggingssporsmal/formVariants/formVariantConfigs/flervalgV1Config.ts
@@ -1,7 +1,7 @@
 import z from "zod";
 import type { KartleggingsspormalFormFieldId } from "../../fieldDefinitions/KartleggingsspormalFormFieldId";
 import { fieldSchemas } from "../../fieldSchemas/fieldSchemas";
-import { defineVariantConfig } from "./formVariantConfigs";
+import { defineVariantConfig } from "../types/FormVariantConfig";
 
 export const flervalgV1Config = defineVariantConfig({
   fieldsInOrder: [
