+++ src/forms/kartleggingssporsmal/formVariants/formVariants.ts
@@ -1,4 +1,6 @@
-import z from "zod";
+import { flervalgFritekstV1Config } from "./formVariantConfigs/flervalgFritekstV1Config";
+import { flervalgV1Config } from "./formVariantConfigs/flervalgV1Config";
+import type { FormVariant } from "./types/FormVariant";
 
 /**
  * This list must be updated when a new form variant / skjemavariant is added in
@@ -7,7 +9,9 @@ import z from "zod";
  * in the candidate table in `meroppfolging-backend`, and can no longer enter
  * that table).
  */
-const formVariants = ["FLERVALG_V1", "FLERVALG_FRITEKST_V1"] as const;
+export const formVariants = ["FLERVALG_V1", "FLERVALG_FRITEKST_V1"] as const;
 
-export const formVariantSchema = z.enum(formVariants);
-export type FormVariant = z.infer<typeof formVariantSchema>;
+export const formVariantConfigs = {
+  FLERVALG_V1: flervalgV1Config,
+  FLERVALG_FRITEKST_V1: flervalgFritekstV1Config,
+} satisfies Record<FormVariant, unknown>;
