+++ src/forms/kartleggingssporsmal/formVariants/types/FormVariant.ts
@@ -1,0 +1,5 @@
+import z from "zod";
+import { formVariants } from "../formVariants";
+
+export const formVariantSchema = z.enum(formVariants);
+export type FormVariant = z.infer<typeof formVariantSchema>;
